Moved plot._color to utils._color.
