+++ phy/cluster/manual/cluster_metadata.py
@@ -10,7 +10,7 @@ from collections import defaultdict, OrderedDict
 from functools import partial
 from copy import deepcopy
 
-from ...plot._color import _random_color
+from ...utils._color import _random_color
 from ...ext.six import iterkeys, itervalues, iteritems
 from ._utils import _unique, _spikes_in_clusters
 from ._update_info import UpdateInfo
