+++ phy/plot/tests/test_waveforms.py
@@ -12,7 +12,7 @@ from vispy import app
 from ...cluster.manual.cluster_metadata import ClusterMetadata
 from ...utils.logging import set_level
 from ..waveforms import Waveforms, WaveformView
-from .._color import _random_colors
+from ...utils._color import _random_colors
 from ...datasets.mock import (artificial_waveforms, artificial_masks,
                               artificial_spike_clusters)
 from ...electrode.mea import staggered_positions
